refactor(post): restructure post module into multiple folders for better organization
