+++ src/components/posts/components/post/index.ts
@@ -1,0 +1,3 @@
+export { default as PostContent } from "./postContent/PostContent";
+export { default as PostFooter } from "./postFooter/PostFooter";
+export { default as PostHeader } from "./postHeader/PostHeader";
