+++ src/components/posts/components/post/postFooter/PostFooter.tsx
